Refactored code into packages removed test files for publishing renamed code for self documenting code
