--- src/main/java/SecretWords.java
@@ -1,13 +1,0 @@
-public class SecretWords {
-
-  private final WordList list;
-
-  public SecretWords(WordList list) {
-    this.list = list;
-    // TODO: 10/4/2022 Output secret words from enum class.
-  }
-
-  public WordList list() {
-    return list;
-  }
-}
