+++ src/main/java/com/teamproject/pigsandbulls/controller/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Base package for classes and making up the implementation of Pigs And Bulls.
+ * Any class located directly in this package is an entry point suitable for launch as a Java
+ * application&mdash; such a class has a {@code public static void main(String[] args)} method.
+ */
+package com.teamproject.pigsandbulls.controller;
