+++ src/main/java/com/teamproject/pigsandbulls/model/UserInputValidator.java
@@ -1,0 +1,24 @@
+package com.teamproject.pigsandbulls.model;
+
+import com.teamproject.pigsandbulls.controller.PigsBulls;
+
+public class UserInputValidator {
+
+  public void checkForCharacterInput(String userGuess) {
+    if (!userGuess.matches("^[A-Z]{5}$") && userGuess.length() == PigsBulls.NUMBER_OF_CHARACTERS) {
+      throw new IllegalArgumentException("You can provide only characters!");
+    }
+  }
+
+  public void checkOversizedWord(String userGuess) {
+    if (userGuess.length() > PigsBulls.NUMBER_OF_CHARACTERS) {
+      throw new IllegalArgumentException("You can provide only 5 characters!");
+    }
+  }
+
+  public void checkUndersizedWord(String userGuess) {
+    if (userGuess.length() < PigsBulls.NUMBER_OF_CHARACTERS) {
+      throw new IllegalArgumentException("You must provide at least 5 characters!");
+    }
+  }
+}
