+++ src/main/java/com/teamproject/pigsandbulls/model/WordList.java
@@ -1,3 +1,5 @@
+package com.teamproject.pigsandbulls.model;
+
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -7,10 +9,6 @@ import java.util.List;
 import java.util.Random;
 import java.util.stream.Collectors;
 
-//overhaul. file for 5 letter words online. class that contains list of strings that are in the file. the class could read those word in its constructor.
-//read in from a file
-//shuffle the list of words and then iterate
-//field that is a list of string and is populated from a list of words read from a file
 public class WordList {
 
   private final List<String> words;
