+++ src/main/java/com/teamproject/pigsandbulls/model/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Classes responsible for managing the internal system state, including validation
+ * and processing of actions in the game that may result in state changes.
+ */
+package com.teamproject.pigsandbulls.model;
