+++ src/main/java/com/teamproject/pigsandbulls/view/OutputMessages.java
@@ -1,12 +1,14 @@
+package com.teamproject.pigsandbulls.view;
+
 import java.util.Scanner;
 
 public class OutputMessages {
 
-  void displayLogo() {
+  public void displayTitle() {
     System.out.println("PIGS AND BULLS \n");
   }
 
-  void displayInstructions(){
+  public void displayInstructions() {
     System.out.println("Welcome to the pigs and bulls game \n" +
         "Your characters need to be all different: [FRUIT] \n" +
         "The program will give you the number of matches. \n" +
@@ -18,39 +20,33 @@ public class OutputMessages {
         " You will have 10 guesses to find the secret word. Good Luck! \n");
   }
 
-  String inputUserName() {
+  public String inputUserName() {
     System.out.println("Enter a username:");
     Scanner keyboardInput = new Scanner(System.in);
     return keyboardInput.nextLine();
   }
 
-  void displayTheEnd(String userName) {
-    System.out.println("Congratulations " + userName + "!" + " You guessed the code!");
-  }
-
-  void displayPlayAgain(String userName){
-    System.out.println("Hey " + userName + "!" + " Would you like to play again?");
-  }
-
-  void displayMessageProvideFiveCharacters(){
+  public void askUserForGuess() {
     System.out.println("Please enter a 5 letter word");
   }
 
-  void displayTurns(int turnCounter) {
+  public void displayTurns(int turnCounter) {
     System.out.print("Turn: " + turnCounter + "/10: ");
   }
 
-  void displayBullsAndPigs(int bulls, int pigs){
+  public void displayBullsAndPigs(int bulls, int pigs) {
     System.out.println(" Bull(s): " + bulls + " Pig(s): " + pigs);
   }
 
-  void displayNoMoreTurnsLeft(String userName) {
-    System.out.println("Sorry, " + userName  + ", there are no more attempts left: game ended!");
+  public void displayNoMoreTurnsLeft(String userName) {
+    System.out.println("Sorry, " + userName + ", there are no more attempts left: game ended!");
   }
 
-  void displaySecret(String secret) {
+  public void displaySecret(String secret) {
     System.out.println("The secret word was: " + secret);
   }
 
-
+  public void displayWinner(String userName) {
+    System.out.println("Congratulations " + userName + "!" + " You guessed the code!");
+  }
 }
