+++ src/main/java/com/teamproject/pigsandbulls/view/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Classes are responsible for constructing text-based representations of the system
+ * state and key transitions to the console (user).
+ */
+package com.teamproject.pigsandbulls.view;
