+++ src/main/resources/banner.txt
@@ -1,0 +1,8 @@
+  _____ _                             _   ____        _ _     
+ |  __ (_)                           | | |  _ \      | | |    
+ | |__) _  __ _ ___    __ _ _ __   __| | | |_) |_   _| | |___ 
+ |  ___| |/ _` / __|  / _` | '_ \ / _` | |  _ <| | | | | / __|
+ | |   | | (_| \__ \ | (_| | | | | (_| | | |_) | |_| | | \__ \
+ |_|   |_|\__, |___/  \__,_|_| |_|\__,_| |____/ \__,_|_|_|___/
+           __/ |                                              
+          |___/
