--- src/test/java/PigsBullsTest.java
@@ -1,39 +1,0 @@
-import static org.junit.jupiter.api.Assertions.*;
-
-import java.util.Scanner;
-import org.junit.jupiter.api.Test;
-
-class PigsBullsTest {
-
-  @Test
-  void countBullsAndPigs() {
-    String msg = "WHILE";
-    assertEquals(5, msg.length());
-    assertTrue(5 == msg.length());
-  }
-
-  @Test
-  void secretWord() {
-    int expected = 5;
-    public class testCountBullsAndPigs() {
-    Scanner scanner = new Scanner(System.in);
-    String msg = "WHILE";
-
-
-    // TODO Create test program edge cases (i.e - < or > than 5 letters, double letters, symbols, following ASCII - Uppercase: 65 to 90; Lowercase: 97 to 122).
-
-    // Todo Create test for secret word
-
-    // Todo create test for bull(s) numbers
-
-    // Todo create test for pig(s) numbers
-
-    // Todo create test for letter casing (i.e - UPPER vs lower or does it matter)
-
-    // Todo Finish the single player version of game.
-
-    // Todo create Enum dictionary
-
-    // Todo TBD
-  }
-}
