--- src/test/java/WordListTest.java
@@ -1,5 +1,0 @@
-import static org.junit.jupiter.api.Assertions.*;
-
-class WordListTest {
-
-}
